[webdriver-client] Move transport class for WebDriver BiDi into its own module.
Differential Revision: https://phabricator.services.mozilla.com/D131605

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1693839
gecko-commit: 95d37ec74f93824d4990555c83055bc175ebd80f
gecko-reviewers: webdriver-reviewers, jdescottes
